backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.5",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyjwt>=2.11.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
